chore: move hooks to separate files
